pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-obsidianRAG"
version = "0.1.0"
description = "MCP server for semantic search over Obsidian notes using local RAG"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Ernest Koe" }]
dependencies = [
    "chromadb>=0.4.0",
    "watchdog>=3.0.0",
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "openai>=1.0.0",
    "platformdirs>=3.0.0",
    "tomli-w>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/ernestkoe/mcp-obsidianRAG"
Repository = "https://github.com/ernestkoe/mcp-obsidianRAG"
Issues = "https://github.com/ernestkoe/mcp-obsidianRAG/issues"

[project.scripts]
obsidian-rag = "obsidian_rag.cli:main"
obsidian-rag-mcp = "obsidian_rag.server:run_server"

[tool.obsidian-rag]
vault_path = "/Users/ernestkoe/Documents/Brave Robot"
data_path = "./data"
ollama_url = "http://localhost:11434"
embedding_model = "nomic-embed-text"
exclude = ["attachments/**", ".obsidian/**", ".trash/**"]

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_rag"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pyright>=1.1.0",
]
